+++ flake.nix
@@ -1,5 +1,5 @@
 {
-  description = "plutarch-script-export";
+  description = "liqwid-script-export";
 
   inputs = {
     nixpkgs.follows = "plutarch/nixpkgs";
@@ -23,7 +23,12 @@
         "plutarch/haskell-nix/nixpkgs-unstable";
     };
 
+    ply.url = "github:mlabs-haskell/ply?ref=master";
+
     liqwid-nix.url = "github:Liqwid-Labs/liqwid-nix";
+    liqwid-plutarch-extra.url = "github:Liqwid-Labs/liqwid-plutarch-extra";
+    plutarch-numeric.url =
+      "github:liqwid-labs/plutarch-numeric?ref=main";
   };
 
   outputs = inputs@{ liqwid-nix, ... }:
@@ -35,6 +40,12 @@
       [
         liqwid-nix.haskellProject
         liqwid-nix.plutarchProject
+        (liqwid-nix.addDependencies [
+          "${inputs.ply}/ply-core"
+          "${inputs.ply}/ply-plutarch"
+          "${inputs.liqwid-plutarch-extra}"
+          "${inputs.plutarch-numeric}"
+        ])
         (liqwid-nix.enableFormatCheck [
           "-XQuasiQuotes"
           "-XTemplateHaskell"