+++ src/ScriptExport/ScriptInfo.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE UndecidableInstances #-}
+
 {- |
 Module     : Agora.ScriptInfo
 Maintainer : emi@haskell.fyi
@@ -8,6 +13,15 @@ Exportable script bundles for off-chain consumption.
 module ScriptExport.ScriptInfo (
   -- * Types
   ScriptInfo (..),
+  Linker (..),
+  RawScriptExport (..),
+  ScriptExport (..),
+
+  -- * Linker utilities
+  runLinker,
+  fetchTS,
+  getParam,
+  getRawScripts,
 
   -- * Introduction functions
   mkScriptInfo,
@@ -16,18 +30,234 @@ module ScriptExport.ScriptInfo (
   mkStakeValidatorInfo,
 ) where
 
-import Aeson.Orphans ()
 import Cardano.Binary qualified as CBOR
 import Codec.Serialise qualified as Codec
+import Control.Monad.Except (Except, MonadError, runExcept, throwError)
+import Control.Monad.Reader (MonadReader, ReaderT, ask, asks, runReaderT)
 import Data.Aeson qualified as Aeson
+import Data.Aeson.Types qualified as Aeson
+import Data.ByteString qualified as BS
 import Data.ByteString.Base16 qualified as Base16
 import Data.ByteString.Lazy qualified as LBS
 import Data.ByteString.Short qualified as SBS
-import Data.Text (Text)
+import Data.Kind (Type)
+import Data.Map (Map, (!?))
+import Data.Text (Text, pack)
+import Data.Text.Encoding qualified as Text
 import GHC.Generics qualified as GHC
-import Plutarch (ClosedTerm, Config (Config, tracingMode), TracingMode (NoTracing))
-import Plutarch.Api.V2 (PMintingPolicy, PStakeValidator, PValidator, mkMintingPolicy, mkStakeValidator, mkValidator, scriptHash)
-import PlutusLedgerApi.V2 (MintingPolicy (getMintingPolicy), Script, ScriptHash, StakeValidator (getStakeValidator), Validator (getValidator))
+import Optics (view)
+import Optics.TH (makeFieldLabelsNoPrefix)
+import Plutarch (
+  ClosedTerm,
+  Config (Config, tracingMode),
+  TracingMode (NoTracing),
+ )
+import Plutarch.Api.V2 (
+  PMintingPolicy,
+  PStakeValidator,
+  PValidator,
+  mkMintingPolicy,
+  mkStakeValidator,
+  mkValidator,
+  scriptHash,
+ )
+import Plutarch.Orphans ()
+import PlutusLedgerApi.V2 (
+  MintingPolicy (getMintingPolicy),
+  Script,
+  StakeValidator (getStakeValidator),
+  Validator (getValidator),
+ )
+import Ply (ScriptRole, TypedScript, TypedScriptEnvelope)
+import Ply.Core.TypedReader (TypedReader, mkTypedScript)
+
+{- | Type for holding parameterized scripts. This type represents the
+     raw scripts which needs to be linked in order to be deployed.
+
+ @since 2.0.0
+-}
+newtype RawScriptExport = RawScriptExport
+  { rawScripts :: Map Text TypedScriptEnvelope
+  }
+  deriving stock
+    ( -- | @since 2.0.0
+      Show
+    , -- | @since 2.0.0
+      Eq
+    , -- | @since 2.0.0
+      GHC.Generic
+    )
+  deriving anyclass
+    ( -- | @since 2.0.0
+      Aeson.ToJSON
+    , -- | @since 2.0.0
+      Aeson.FromJSON
+    )
+
+{- | Type for holding ready-to-go scripts. Ready-to-go scripts are
+     scripts that can be used on-onchain without any extra
+     arguments. They are `MintingPolicy ~ StakingValidator ~ PData ->
+     PScriptContext -> POpaque` and `Validator ~ PData -> PData ->
+     PScriptContext -> POpaque`.
+
+ @since 2.0.0
+-}
+data ScriptExport a = ScriptExport
+  { scripts :: Map Text Script
+  , information :: a
+  }
+  deriving stock
+    ( -- | @since 2.0.0
+      Show
+    , -- | @since 2.0.0
+      Eq
+    , -- | @since 2.0.0
+      GHC.Generic
+    )
+
+-- | @since 2.0.0
+data LinkerError
+  = VersionMismatch
+  | ScriptNotFound Text
+  | ScriptTypeMismatch Text
+  | Other Text
+  deriving stock
+    ( -- | @since 2.0.0
+      Show
+    , -- | @since 2.0.0
+      Eq
+    , -- | @since 2.0.0
+      GHC.Generic
+    )
+  deriving anyclass
+    ( -- | @since 2.0.0
+      Aeson.ToJSON
+    , -- | @since 2.0.0
+      Aeson.FromJSON
+    )
+
+{- | Type for linker that makes ready-to-go script from parameter and
+     raw script.
+
+ @since 2.0.0
+-}
+newtype Linker param a
+  = Linker (ReaderT (RawScriptExport, param) (Except LinkerError) a)
+  deriving
+    ( -- | @since 2.0.0
+      Functor
+    , -- | @since 2.0.0
+      Applicative
+    , -- | @since 2.0.0
+      Monad
+    , -- | @since 2.0.0
+      MonadError LinkerError
+    , -- | @since 2.0.0
+      MonadReader (RawScriptExport, param)
+    )
+    via (ReaderT (RawScriptExport, param) (Except LinkerError))
+
+{- | Run linker with parameter and raw scripts.
+
+ @since 2.0.0
+-}
+runLinker ::
+  forall (param :: Type) (a :: Type).
+  Linker param (ScriptExport a) ->
+  RawScriptExport ->
+  param ->
+  Either LinkerError (ScriptExport a)
+runLinker (Linker linker) rs p =
+  runExcept $ runReaderT linker (rs, p)
+
+-- | @since 2.0.0
+fetchTS ::
+  forall (rl :: ScriptRole) (params :: [Type]) (lparam :: Type).
+  TypedReader rl params =>
+  Text ->
+  Linker lparam (TypedScript rl params)
+fetchTS t = do
+  (scr, _) <- ask
+  case view #rawScripts scr !? t of
+    Just scr' ->
+      case mkTypedScript @rl @params scr' of
+        Right x -> return x
+        Left err -> throwError $ ScriptTypeMismatch $ pack $ show err
+    Nothing -> throwError $ ScriptNotFound t
+
+-- | @since 2.0.0
+getParam :: forall (lparam :: Type). Linker lparam lparam
+getParam = asks snd
+
+-- | @since 2.0.0
+getRawScripts :: forall (lparam :: Type). Linker lparam RawScriptExport
+getRawScripts = asks fst
+
+-- Internal helper for CBOR deserialization
+newtype CBORSerializedScript = CBORSerializedScript SBS.ShortByteString
+
+instance CBOR.FromCBOR CBORSerializedScript where
+  fromCBOR = CBORSerializedScript . SBS.toShort <$> CBOR.fromCBOR
+
+cborToScript :: BS.ByteString -> Either CBOR.DecoderError Script
+cborToScript x =
+  Codec.deserialise
+    . LBS.fromStrict
+    . SBS.fromShort
+    . (\(CBORSerializedScript x) -> x)
+    <$> CBOR.decodeFull' x
+
+-- | @since 2.0.0
+instance
+  forall (a :: Type).
+  Aeson.ToJSON a =>
+  Aeson.ToJSON (ScriptExport a)
+  where
+  toJSON (ScriptExport scripts info) =
+    Aeson.toJSON $
+      Aeson.object
+        [ "info" Aeson..= info
+        , "scripts" Aeson..= (toJSONScript <$> scripts)
+        ]
+    where
+      toJSONScript scr =
+        let raw = LBS.toStrict $ Codec.serialise scr
+            cbor = CBOR.serialize' $ SBS.toShort raw
+         in Aeson.toJSON $
+              Aeson.object
+                [ "cborHex" Aeson..= Base16.encodeBase16 cbor
+                , "rawHex" Aeson..= Base16.encodeBase16 raw
+                , "hash" Aeson..= scriptHash scr
+                ]
+
+-- | @since 2.0.0
+instance
+  forall (a :: Type).
+  Aeson.FromJSON a =>
+  Aeson.FromJSON (ScriptExport a)
+  where
+  parseJSON (Aeson.Object v) =
+    ScriptExport
+      <$> (v Aeson..: "scripts" >>= Aeson.liftParseJSON parseJSONScriptMap Aeson.parseJSON)
+      <*> v Aeson..: "info"
+    where
+      parseAndDeserialize v =
+        Aeson.parseJSON v
+          >>= either (fail . show) (either (fail . show) pure . cborToScript)
+            . Base16.decodeBase16
+            . Text.encodeUtf8
+
+      parseJSONScriptMap :: Aeson.Value -> Aeson.Parser Script
+      parseJSONScriptMap (Aeson.Object v) =
+        v Aeson..: "cborHex" >>= parseAndDeserialize
+      parseJSONScriptMap invalid =
+        Aeson.prependFailure
+          "parsing Script from ScriptExport failed, "
+          (Aeson.typeMismatch "Object" invalid)
+  parseJSON invalid =
+    Aeson.prependFailure
+      "parsing ScriptExport failed, "
+      (Aeson.typeMismatch "Object" invalid)
 
 {- | Bundle containing a 'Validator' and its hash.
 
@@ -38,8 +268,6 @@ data ScriptInfo = ScriptInfo
   -- ^ The validator script encoded as cbor hex.
   , rawHex :: Text
   -- ^ The validator script encoded as raw hex.
-  , hash :: ScriptHash
-  -- ^ Hash of the validator.
   }
   deriving stock
     ( -- | @since 1.1.0
@@ -63,7 +291,6 @@ mkScriptInfo script =
    in ScriptInfo
         { cborHex = Base16.encodeBase16 scriptCBOR
         , rawHex = Base16.encodeBase16 scriptRaw
-        , hash = scriptHash script
         }
 
 exportConfig :: Config
@@ -95,3 +322,12 @@ mkValidatorInfo term =
 mkStakeValidatorInfo :: ClosedTerm PStakeValidator -> ScriptInfo
 mkStakeValidatorInfo term =
   mkScriptInfo (getStakeValidator $ mkStakeValidator exportConfig term)
+
+----------------------------------------
+-- Field Labels
+
+-- | @since 2.0.0
+makeFieldLabelsNoPrefix ''ScriptExport
+
+-- | @since 2.0.0
+makeFieldLabelsNoPrefix ''RawScriptExport