+++ src/ScriptExport/Types.hs
@@ -1,3 +1,7 @@
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE UndecidableInstances #-}
+
 {- |
 Module     : ScriptExport.Types
 Maintainer : emi@haskell.fyi
@@ -6,22 +10,35 @@ Description: Param and script types for generation.
 Param and script types for generation.
 -}
 module ScriptExport.Types (
+  ServeElement (..),
   ScriptQuery (..),
   Builders,
+  handleServe,
+  getBuilders,
   runQuery,
   insertBuilder,
+  insertStaticBuilder,
+  insertScriptExportWithLinker,
   toList,
 ) where
 
+import Control.Monad.Except
 import Data.Aeson qualified as Aeson
 import Data.ByteString.Lazy.Char8 qualified as LBS
-import Data.Coerce (coerce)
 import Data.Default.Class (Default (def))
 import Data.Hashable (Hashable)
+import Data.Kind (Type)
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
-import Data.Text (Text)
+import Data.Text (Text, pack, unpack)
 import GHC.Generics qualified as GHC
+import Optics.TH (makeFieldLabelsNoPrefix)
+import ScriptExport.ScriptInfo (
+  Linker,
+  RawScriptExport,
+  ScriptExport,
+  runLinker,
+ )
 import Servant qualified
 
 {- | Query data for getting script info.
@@ -30,7 +47,7 @@ import Servant qualified
 -}
 data ScriptQuery = ScriptQuery
   { name :: Text
-  , paramsPayload :: Aeson.Value
+  , param :: Maybe Aeson.Value
   }
   deriving anyclass
     ( -- | @since 1.0.0
@@ -58,55 +75,135 @@ data ScriptQuery = ScriptQuery
      @since 1.0.0
 -}
 runQuery :: ScriptQuery -> Builders -> Servant.Handler Aeson.Value
-runQuery s =
+runQuery (ScriptQuery name param) =
   maybe
     (Servant.throwError Servant.err404 {Servant.errBody = "Builder not found"})
-    ($ s.paramsPayload)
-    . Map.lookup s.name
+    (toServantErr . runExcept . handleServe param)
+    . Map.lookup name
     . getBuilders
+  where
+    toServantErr (Left err) =
+      Servant.throwError
+        Servant.err400
+          { Servant.errBody = (LBS.pack . unpack) err
+          }
+    toServantErr (Right x) = pure x
+
+{- | Possible data to request.
+
+     @since 2.0.0
+-}
+data ServeElement where
+  ServeRawScriptExport ::
+    forall (a :: Type) (param :: Type).
+    (Aeson.FromJSON param, Aeson.ToJSON a) =>
+    RawScriptExport ->
+    Linker param (ScriptExport a) ->
+    ServeElement
+  ServeJSON ::
+    forall (s :: Type).
+    (Aeson.ToJSON s) =>
+    s ->
+    ServeElement
+  ServeJSONWithParam ::
+    forall (p :: Type) (s :: Type).
+    (Aeson.FromJSON p, Aeson.ToJSON s) =>
+    (p -> s) ->
+    ServeElement
+
+{- | Handle `ServeElement` and returns JSON.
+
+     @since 2.0.0
+-}
+handleServe :: Maybe Aeson.Value -> ServeElement -> Except Text Aeson.Value
+handleServe _ (ServeJSON x) = pure $ Aeson.toJSON x
+handleServe (Just arg) (ServeJSONWithParam f) =
+  case Aeson.fromJSON arg of
+    Aeson.Error e ->
+      throwError $ pack e
+    Aeson.Success v' ->
+      pure . Aeson.toJSON $ f v'
+handleServe (Just arg) (ServeRawScriptExport scr linker) =
+  case Aeson.fromJSON arg of
+    Aeson.Error e ->
+      throwError $ pack e
+    Aeson.Success v' ->
+      case runLinker linker scr v' of
+        Left e -> throwError . pack . show $ e
+        Right x -> pure . Aeson.toJSON $ x
+handleServe Nothing (ServeRawScriptExport scr _) = pure $ Aeson.toJSON scr
+handleServe Nothing (ServeJSONWithParam _) = throwError "Query expects an argument, but nothing is given"
 
 {- | Represents a list of named pure functions.
 
-     @since 1.0.0
+     @since 2.0.0
 -}
-newtype Builders = Builders
-  { getBuilders :: Map Text (Aeson.Value -> Servant.Handler Aeson.Value)
-  }
+newtype Builders
+  = Builders (Map Text ServeElement)
+  deriving
+    ( -- | @since 1.0.0
+      Semigroup
+    , -- | @since 1.0.0
+      Monoid
+    )
+    via (Map Text ServeElement)
 
--- | @since 1.0.0
+-- | @since 2.0.0
+getBuilders ::
+  Builders ->
+  Map Text ServeElement
+getBuilders (Builders b) = b
+
+{- | Get a list of the available builders.
+
+ @since 2.0.0
+-}
+toList :: Builders -> [Text]
+toList = Map.keys . getBuilders
+
+-- | @since 2.0.0
 instance Default Builders where
   def = Builders Map.empty
 
 {- | Insert a pure function into the Builders map.
 
-     @since 1.0.0
+     @since 2.0.0
 -}
 insertBuilder ::
-  forall p s.
+  forall (p :: Type) (s :: Type).
   (Aeson.FromJSON p, Aeson.ToJSON s) =>
   Text ->
   (p -> s) ->
-  Builders ->
   Builders
-insertBuilder k = coerce . Map.insert k . throughJSON
-  where
-    throughJSON ::
-      forall p s.
-      (Aeson.FromJSON p, Aeson.ToJSON s) =>
-      (p -> s) ->
-      (Aeson.Value -> Servant.Handler Aeson.Value)
-    throughJSON f v =
-      case Aeson.fromJSON v of
-        Aeson.Error e ->
-          Servant.throwError $
-            Servant.err400
-              { Servant.errBody = LBS.pack e
-              }
-        Aeson.Success v' -> pure . Aeson.toJSON $ f v'
+insertBuilder k f =
+  Builders $ Map.insert k (ServeJSONWithParam f) mempty
 
-{- | Get a list of the available builders.
+insertStaticBuilder ::
+  forall (a :: Type).
+  (Aeson.ToJSON a) =>
+  Text ->
+  a ->
+  Builders
+insertStaticBuilder k x =
+  Builders $ Map.insert k (ServeJSON x) mempty
 
-     @since 1.0.0
+{- | Insert a 'RawScriptExport' and 'ScriptLinker' to the Builders Map. The
+     builder will return applied `ScriptExport` with given parameter.
+
+     @since 2.0.0
 -}
-toList :: Builders -> [Text]
-toList = Map.keys . getBuilders
+insertScriptExportWithLinker ::
+  forall (param :: Type) (a :: Type).
+  (Aeson.FromJSON param, Aeson.ToJSON a) =>
+  Text ->
+  RawScriptExport ->
+  Linker param (ScriptExport a) ->
+  Builders
+insertScriptExportWithLinker k scr linker =
+  Builders $ Map.insert k (ServeRawScriptExport scr linker) mempty
+
+----------------------------------------
+-- Field Labels
+
+-- | @since 2.0.0
+makeFieldLabelsNoPrefix ''ScriptQuery