+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 2.0.0 -- 2022-10-12
+
+### Added
+
+- `ScriptExport`, `RawScriptExport`, and `Linker` provides new standardized data structure for script exportation.
+- Utility functions for new `ScriptExport` types.
+- `insertStaticBuilder`, and `insertScriptExportWithLinker` was added.
+- `exportMain` is new entry point for the script exporting program.
+- `exportFile` exports scripts to file.
+- Other internal data structure changed to better support new `ScriptExport` types.
+
 ## 1.1.1 -- 2022-09-15
 
 ### Added